Integrate MessageQueueService with database

+++ BillboardServer/Repositories/BillboardRepository.cs
@@ -1,33 +1,42 @@
 ﻿using BillboardServer.Data;
 using BillboardServer.Models;
+using BillboardServer.Extensions;
 
 using Microsoft.EntityFrameworkCore;
 
-namespace BillboardServer.Repositories
+namespace BillboardServer.Repositories;
+
+public class BillboardRepository
 {
-    public class BillboardRepository
+    private readonly BillboardDbContext _context;
+
+    public BillboardRepository(BillboardDbContext context)
     {
-        private readonly BillboardDbContext _context;
+        _context = context;
+    }
 
-        public BillboardRepository(BillboardDbContext context)
+    public async Task<IEnumerable<Message>> GetAllMessagesAsync()
+    {
+        var allMessages = await _context.Messages
+            .OrderBy(m => m.CreatedAt)
+            .ToListAsync();
+
+        if (allMessages.Any())
         {
-            _context = context;
-        }
-
-        public async Task<IEnumerable<Message>> GetAllMessagesAsync()
-        {
-            var allMessages = await _context.Messages.OrderBy(m => m.CreatedAt).ToListAsync();
-
+            allMessages.NormalizeDates();
             _context.Messages.RemoveRange(allMessages);
             await _context.SaveChangesAsync();
-
-            return allMessages;
         }
 
+        return allMessages;
+    }
 
-        public async Task AddMessagesAsync(IEnumerable<Message> messages)
+    public async Task AddMessagesAsync(IEnumerable<Message> messages)
+    {
+        if (messages.Any())
         {
-            await _context.Messages.AddRangeAsync(messages);
+            messages.NormalizeDates();
+            _context.Messages.AddRange(messages);
             await _context.SaveChangesAsync();
         }
     }
